Completely Overhauled OreCrusherContainer class. Completely Overhauled OreCrusherTileEntity class.
Added OreCrusherStateData to keep track of state information as well as read/write NBT StateData.

Main OreOverhaul Class is still messy but it is registry events are going to be moved to respective packages instead of being contained in the main file.

Fixed bug where OreCrusher would open GUI but also allow you to place a block